Screen Share Experiment
This experiments looks at a couple of factors that may impact some
screen share content.

In VBR tests increasing the minimum boost on key frames seemed to help
quite a lot on a subsection of screen content such as the Wikipedia
clip where the current code allocates only a small number of bits to
the key frame (much less than to ARF frames. This should not impact
fixed Q encodes though.

In some local VBR testing reducing the RD multiplier also seemed to
help. The rationale here is that by so doing you direct more bits into
the prediction part of the signal. In the case of this type of screen
share content it is sometimes possible to get very good predictors
but this may mean testing modes that might otherwise be skipped.

The final change reduces the threshold for doing exhaustive or mesh
motion searches. The rationale here is that for dense moving text on
clips such as in the Wikipedia clip there may be an increased risk of
the step search finding a false minimum an an exhaustive or mesh
search may be able to find a much better prediction.

The RD multiplier change did not seem to provide a gain when I did
local 1 pass fixed Q tests in the range Q40 - Q46. Theses runs gave
a higher data rate than my previous VBR tests so it may be that there
is some Q dependency here. However, for now this change is commented
out.

Lowering the exhaustive motion search though gave gains in the order
of 4-6%

At the moment the changes will only be active when the command line
flag for screen content is explicitly set.

This is an experimental patch so should not be submitted yet.

Change-Id: I6bcb7b6d5fb8a74f9e0c4b2fa31941156549e19a