Clean up redundant functions in temporal filtering
This CL cleans up the redundant functions and unit tests after the
recent unification of the two filtering strategies. In particular, YUV
filtering strategy is removed and the filtering process is simplified.

NOTE: The encoding speed is not affacted after the unification.

See https://aomedia-review.googlesource.com/c/aom/+/104826 for more
details on the strategy unifcation.

Change-Id: Idf9775a397dd97d36d35309c718c4c1b2f041661