separate compound/singleref convolve functions
Instead of indirectly referencing them via the sf->convolve
map of 3 parameters to function, directly reference them and
clean up unused parameters.

This change also removes one redundant set of convolve
functions: the av1_convolve_2d_copy in av1. It also enables
optimizations for MIPs architectures for this function (which
were present but not enabled). Other redundant convolve
functions will be removed in other changes.

The convolve testing framework has been reworked; the time
to run the tests has been reduced slightly; it should also
be clearer.

First part of a clean-up to remove conv_params->dst.

This is a similar change as
https://aomedia-review.googlesource.com/c/aom/+/107922
but for the master branch.

BUG=aomedia:2634

Change-Id: I515526437e9dc40db3b54a264c8cde416e3a1e4c