separate compound/singleref convolve functions
First part of a clean-up to remove conv_params->dst.
This is a similar change as
https://aomedia-review.googlesource.com/c/aom/+/107922
but for the master branch.

1. Directly reference convolves and clean-up unused parameters
(instead of the indirect sf->convolve map)

2. Remove av1_convolve_2d_copy - redundant with
aom_convolve_copy. Other redundant functions will be removed
in future changes.

3. Enable MIPS optimizations for aom_convolve_copy -- they
are already present, but were not enabled.

4. Re-work the convolve testing framework. The time
to run the tests has been reduced by ~7%; the code is less
compact but I believe it is easier to read.

BUG=aomedia:2634

Change-Id: I515526437e9dc40db3b54a264c8cde416e3a1e4c