Make function declaration and definition match.
The 7th parameter of av1_make_default_fullpel_ms_params(), int
fine_search_interval, must have the same type, including constness, in
the declaration and definition. Otherwise Visual C++ warns about this.

Note: In C++ it is okay to add 'const' to parameters in function
definitions. In C, gcc and clang allow this as a language extension, but
Visual C++ (when compiling C code) does not allow this.

Change-Id: Ia15fce7137494055f7d8e61ad9593c141b5d72de