Remove AOM_EMULATE_INTTYPES
Why is it safe to remove AOM_EMULATE_INTTYPES?

The #if defined(AOM_EMULATE_INTTYPES) block does not define the int64_t
and uint64_t types, but uint64_t is used later in this file. This
implies the #if defined(AOM_EMULATE_INTTYPES) block must be unused and
therefore can be removed.

A second reason is that this file includes <inttypes.h>. If <inttypes.h>
is available, then <stdint.h> must also be available.

Also remove the #if defined(_MSC_VER) && _MSC_VER < 1800 block. 1800 is
Visual Studio 2013 (12.0). libaom requires Visual Studio 2017 (15.0) or
later.

Change-Id: I217f4b883fe16fee071c0a65dfcdd50bec382483