Eliminate duplicate highbd-convolve-copy
aom_highbd_convolve_copy and av1_highbd_convolve_copy_sr are nearly
identical (the latter supports widths of 2, whereas the former doesn't,
but otherwise the code is the same). Standardize on the latter,
and put it in the aom_dsp directory where it can be shared with other
codecs like av2.

Change-Id: I63aebfea71a4ca7cd164508fbfaee8a4985ae3c2