Fix a frame_type check using outdated values
Currently, when the encoder tries to check for screen_content decision
in the second pass, it checks whether current_frame's frame_type is
INTRA_FRAME or KEY_FRAME. However, the check is made before the current
frame_type is finalized, so the encoder could make screen content in
INTER_FRAME as well.

This CL moves the screen_content decision to a later point in the
encoding process and updates the frame_type slightly earlier.

BUG=aomedia:2637

STATS_CHANGED

Change-Id: Iacc0bd96bc872c9afa6507d6923e00d1263d728c