Adopt new strategy for allocating ARF group bits as standard.
This patch takes a new and simpler strategy for allocating ARF bits out
from under a flag after testing of the trade off of increased rate accuracy
and metrics.

The old strategy in essence calculated a boost score for the ARF group that
determined the proportion of the total bits allocated to the top level ARF.
There was then a new boost calculation for each subsequent ARF layer
to which a proportion of the remaining bits were allocated. This repeated down
to the leaf frames with an ever dwindling pool of bits.  The result was that
the low level ARF frames were usually allocated an unachievably small number
of bits. These frames would then be clamped by the active max Q. If the Q
was allowed to adjust without limits for each frame (e.g. STRICT_RC  flag)
the result was a huge hit on metrics in return for better rate control.

The new strategy uses the top level ARF boost to determine a pool of bits from
the total for the group, that will be allocated as extra bits to ARF frames. The
remaining bits are shared equally among all frames in the group including ARF
and leaf frames. This Insures that even low level ARF frames and leaf frames
get a reasonable minimum allocation of bits (usually similar to the actual
number spent in the previous case given constraints on active min and max q).
The allocation of extra bits to each layer is done according to a much simpler
fixed strategy rather than by computing new boost values.

The behavior when STRICT_RC  is set is hugely improved.
low res -24.355 -27.486 -18.352 -27.127 -29.144

All Results below are (av psnr, ov psnr, ssim, psnr-hvs and vmaf).

Rate control accuracy is given by looking at the number of clips in a set where
accuracy improves vs the number where it gets worse. So for example (rc 25/10)
would mean that 25 clips showed a clear improvement vs 10 that were worse.

Results.
When coding with the overshoot and undershoot % set to 100 (these flags
control a secondary rate control adjustment mechanism), this patch results in
an improvement in rate control accuracy for most clips but a slight penalty
in regard to metrics (except vmaf).

low res 0.689	0.601	1.145	0.703	 0.331	rc 37/0
ugc 360 0.367	0.011	1.029	0.087	-0.484	rc 38/5
mid res 0.752	0.376	1.464	0.392	-0.222	rc 23/2
hd res  0.227	0.117	0.472	0.183	 0.02	rc 32/1

When the overshoot / undershoot % are set to 5 the metrics gap narrows
with overall psnr and vmaf now generally better but the other metrics
(especially ssim) still worse. The ratios in terms of improved rate control
remain similar.

When we then compare a case where the old allocation code uses os/us 90% vs
100% for the new mechanism we see improvements in metrics (except ssim) and
rate accuracy, indicating that the change gives a favorable return in terms of
metrics penalty for improved rate accuracy even without the RC_STRICT flag.

low res -0.114	-0.279	0.698	-0.08	-0.839	rc 29/7
ugc 360 -0.16	-0.769	0.517	-0.489	-0.939	rc 35/14
Mid res and hd res not complete.

STATS_CHANGED

Change-Id: I4ada3d2ca8fa02e3579d660028ecd212a42a4bb7