Update patch set 2
Patch Set 2:

Debargha,
We looked at the quantization operation used in av1_nn_output_prec_reduce, and see that it is rounding the floating point number to a lower precision to avoid the mis-match issues due to the order of addition of float variables. 
In case of first-pass, the precision depends on number of MBs/resolution of the video(i.e., number of additions). Hence we can use either of following approaches.
1. Use precision based on the resolution – Requires tuning for different resolutions
2. Use precision based on the worst case number of MBs – This can reduce the precision overall. 
Also, introducing quantization will change the output of existing single-thread first pass as well, to a lower precision. Due to above, we feel that the approach in the patch/alternatives suggested in the earlier review message are recommended. Please let us know your opinion.

Patch-set: 2