Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Seems like now with overshoot / undershoot % being 100, the results are better (or at least similar) to 90% for previous strategy, in terms of both RC accuracy and metrics. Do you think it could work if we relax the overshoot / undershoot algorithm so that we also get such better results when compared with 100%?

The secondary rc mechanism does not do much with os/us set to 100% but if we move in the direction of some of the changes added in VP9 for corpus vbr mode, which cut off other parts of the clip level feedback loop, we may well see metrics gains. Also we may be able to tune this a bit further even for the 100% os/us setting.

For now though I think the trade off is good, it reduces the dependence on the active Q range and it prevents the allocation algorithm from routinely allocating an un-achievable number of bits to frames.

Patch-set: 2