Modified KF min boost limit.
Add in a minimum KF boost based on the algorithm used to calculate
the arf boost.

This change is specifically targeted at some screen share content where
the KF boost was coming out much less than the first arf boost due
to differences in how boost is calculated in these two cases.

To have the KF quality far below the arf quality generally does not make
sense so this patch uses the function av1_calc_arf_boost() to
set a minimum for the KF.

This results in large gains for the screen share clip Wikipedia.

It has not yet been evaluated for more generally and may need to
be protected under a flag that limits its use to screen share content.

STATS_CHANGED

Change-Id: I74303b49457697d729d8d694ba16a0d388c61866