Update patch set 2
Patch Set 2:

(5 comments)

After investigating the remark about fixing the formatting in the comments, I suspected that even if they were fixed in the comments, the doxygen docs would have the same issue (due to lack of annotation). It turns out this was the case. I ended up doing a full scan through the documentation and fixed everything I could.

Patch-set: 2