Update patch set 2
Patch Set 2: Code-Review-2

These psnr regressions are detailed and explained in the original patch and relate to the fact that it also results in tighter rate control (which almost always causes a regression in raw psnr numbers). E.g.
 
low res 0.689	0.601	1.145	0.703	 0.331	rc 37/0

The patch notes also shows that compared to other mechanisms for tightening up rate accuracy it performs well.  For example, when trying to match rate accuracy with the old code (actually still better with the new)

low res -0.114	-0.279	0.698	-0.08	-0.839	rc 29/7

The old / existing allocation gives an impossible (in many cases) allocation of bits to certain frames and relies on the min and max Q range to prevent severely aberrant behavior. The impact of this change if we actually try and hit the allocated rates for each frame by not imposing Q limits is thus huge. Eg.

Low Res  -24.355 -27.486 -18.352 -27.127 -29.144

As such the old allocation is untenable as a basis for moving forward and trying both to improve rate accuracy and get better vbr metrics.

Future patches will hopefully get back some of the regression but it is also worth noting that even as things stand and with the os and us set to 100% (best metrics with least accurate rate control) VMAF is often better with the new allocation.

Patch-set: 2
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-2