Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This looks like a good catch and a very good outcome. 😊
> 
> Have you checked back to see if the same problem / bug is there in VP9. Obviously this is likely to have much more of an effect in real YT content and not so much in our standard test sets, so could easily have been missed there too.

Just checked VP9 codebase, fortunately it seems that it does not have this problem, since it uses a const pointer + frame offset, instead of changing twopass->stats_in in the input_stats function. Maybe it is better to refactor that part a bit to avoid such issues. I'll see if I could separate the whole twopass structure from the function and hopefully pass in a FIRSTPASS_STATS pointer instead.

Patch-set: 3