Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review+2
> 
> Looks good. Need a rebase. Also, there are no dependencies between the jobs, right?

Hi Yunqing,

When the speed feature 'prune_ref_frame_for_gm_search' is ON, only 2 threads are allowed to work in parallel (one for forward refs and the other for backward refs). As params estimation is confined to a given direction, there is no dependency across threads.

Whereas for boosted frames and for cpu-used 0 (where the speed feature is OFF), more threads can work in parallel. 

Both the scenarios are taken care of while assigning the jobs.

 

We plan to experiment with speed feature being off when more than 2 threads are configured.

Regards,
Remya

Patch-set: 7