Proposed Approach 1
Introduce EXT_WARP config parameter for cases where fractional
offsets are outside [-1,2) range. The warped motion model is
used to obtain the centre pixel of each 4x4 block in the partition
and a translational MC is applied for the 4x4 blocks around the
projected centre using a new 8-tap warp filter.

STATS_CHANGED

Change-Id: I21db2189c12af98a567d72b3ad874737488cef51