Write encoder settings in WebM files
If the output is webm, the command line (minus the binary name,
anything matching the input filename, and output flags) will be
written as an ENCODER_SETTINGS tag in the WebM container.

STATS_CHANGED:
Output file will be slightly larger for WebM files, but underlying
bitstream will be the same.

BUG=aomedia:2593

Change-Id: I03e724d1eea9d6a82ef3068100b7ef96ffb3f87b