Update patch set 7
Patch Set 7:

There is already some adjustment of the filtering based on noise and also based on other factors such as the ARF boost number.  This is interesting because it seems at first sight to go the opposite way in terms of its impact to the existing code which INCREASES filter strength when the clip is noisier. Obviously this is a little different to increasing the number of frames used though. Also there is adjustment to the number of frames based on the ARF boost. I think it might be worth seeing if these adjustments can be merged into one place and also making sure they are not fighting each other. For example the adjustment to the number of frames based on boost may not be needed in addition to this patch and in fact could be fighting it. Look at adjust_arnr_filter() and estimate_strength(). If need b this can be checked in and the adjustments / removal or merger of the other changes can be done as a follow on patch if that is easier.

Patch-set: 7
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>