Add --use-16bit-internal param to aomenc
This parameter is meant to force use of the 16-bit
pipeline when encoding a video even if the source is 8-bit
and 8-bit encoding is desired.

Previously the config flag FORCE_HIGHBITDEPTH_DECODING=1 forced
use of the 16 bit pipeline, but this flag is supposed to control
the decoder side only. The right way to support forcing of
the 16-bit pipeline on the encoder side is with a commandline
parameter.

Change-Id: I9a9c6f3a1d2a0ed4f50f5c91953b321a3a9dfc2e