Update patch set 6
Patch Set 6:

We are seeing some build failures on this patch which we think are unrelated to this change as in the failure logs PSNRs and bitrate values are matching exactly. We are also seeing similar issues on other commits that have been pushed today(For example, https://aomedia-review.googlesource.com/c/aom/+/111941). Please let us know how to take this forward.

Patch-set: 6