Disable WebM encoder settings for debug output
The testing framework assumes that the --debug flag will
create identical WebM containers, even for different builds (as long
as the underlying OBU frames are identical). Since the encoder-settings contain
the build version, it cannot be written out in this case, unless the testing
framework is also changed.

STATS_CHANGED:
WebM files are slightly smaller since the encoder-settings tag
is not written in debug mode (which the testing framework uses).

Change-Id: Icbf57c0913268e9e512d1e3527291fa1d624ea87