Update patch set 12
Patch Set 12:

> Patch Set 11:
> 
> I'd prefer not to add exe files to this repository, documenting the revisions tested and the tools used to build them should enough. Is this something that's going to see continual use or is it something that was a one off? If it's for documentation purposes only maybe a sandbox would be a better place for it. You have permission to create any number of them under refs/heads/sandbox/<username>. Yaowu and Debargha may have other opinions.

Hi, James

thanks for the feedback. I have removed all exe from the bin folder in the latest patch. I also updated the readme file to indicate where to download those tools, or build them. this framework will be used and updated in the future for AV2 common tests, it is not for documentation. AOM test working group has agreed that it should be maintained together with the reference code.

Patch-set: 12