Update patch set 6
Patch Set 6:

Hi Yunqing,
 
Currently in global motion multi-threading, the speed feature 'prune_ref_frame_for_gm_search' limits the parallelism to two-way parallelism.
 
In this patch, we disabled the speed feature to achieve more parallelism when number of threads available are greater than or equal to the maximum number of reference frames allowed for global motion. This change would be helpful for contents with global motion, with very less degradation in speed for non-global-motion contents.

The optimal condition for disabling the speed feature should be based on the actual number of reference frames available for global motion. This needs to be calculated based on
1.	gm_search_type sf
2.	cpi->ref_frame_flags
3.	selective_ref_frame sf / prune_ref_by_selective_ref_frame()

We felt that using 2 and 3 in av1_set_speed_features_framesize_independent() may not be clean. Another option is to reset the sf locally based on above conditions in av1_compute_global_motion_facade() where 1,2 and 3 are available. However we did not prefer this approach as it is not a typical practice. 

Hence, introduced the logic based on 1 only.

The speed feature is disabled in the multi-thread path, only when --row-mt=1 and is unmodified for row-mt=0 to ensure MT bit-match with threads = 1.
In summary,
·        when row-mt == 0, bit matching is verified for threads = 1 and threads = 4 (single-thread vs multi-thread).
·        when row-mt == 1, bit matching is verified for threads = 2 and threads = 4 (multi-thread vs multi-thread).
 
The encode time reduction(in %) seen with the current patch is as below (for 720p 4 threads / 1080p 8 threads):
                       Encode time reduction
Contents               cpu-used 1   cpu-used 2
Global motion             7.46        7.72
Non global motion        -0.29       -0.37
 
Following is the summary of speed-up (in %) for each content:
                              Encode time reduction
Contents                      cpu-used 1   cpu-used 2          
city_720p30_2000kbps           8.35          8.14
city_720p30_600kbps            7.35          7.51
fourpeople_720p60_1600kbps     0.70          0.98
fourpeople_720p60_400kbps      0.59          1.00
night_720p30_2500kbps          0.69          0.09
ped_1080p25_1000kbps          -0.64         -0.17
ped_1080p25_7000kbps          -0.33         -0.33
wikipedia_420_1000kbps         0.76          0.64
blue_sky_1080p30_1000kbps      9.76         10.40
blue_sky_1080p30_7000kbps      4.39          4.85

Note:
-ve indicates encode time increase.
+ve indicates encode time reduction.
 
Please review the patch and merge in case of no issues.

Regards,
Remya

Patch-set: 6
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 17598 <17598@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 9039 <9039@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 13265 <13265@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>