Add NN_RECON experiment
This commit adds NN_RECON experiment.

In this experiment, we add an extra CNN-based reconstruction option at
the partition level. At each partition block, the encoder will signal
whether a cnn-based reconstruction option is on for all the transform
blocks inside. If off, the bitstream with proceed with the usual
dct-based reconstruction. If on, we will first perform a dct-based
reconstruction, then we will apply an cnn-based filter to the result of
dct reconstruction along with some contexts.

This experiment is still work in progress.
Due to complexity consideration, currently this is only applied to
BLOCK_16X16. Furthermore, in the rdopt process, the encoder tests the
cnn filter for all prediction_mode x txfm_type combinations, so the
encoding is very slow (around 600X the baseline speed). We are working
to optimize this rdopt process.

In offline validation outside the encoder, this cnn filter provides
about 0.17 dB gain, but validation on the encoder only shows 0.05%
bitrate reduction on high resolution images. This is mostly due to two
factors:
  1. The bitstream signaling is not very efficient. In particular, it is
  currently independent of the quantization index. Intuitively speaking,
  this reconstruction option should be off for small q-index.
  2. Currently the CNNs are only enabled for TX_16X16 due to high
  complexity cost. Once the rdopt process is sped up more, more CNN
  models can be turned on, giving us more gains.

Note: This commit only adds the necessary bitstream/rdopt changes for
the experiment. The actual model isn't uploaded here.

Change-Id: I650e5d73080f7ac182b390e1d5d2db472f0aeaad