Update patch set 11
Patch Set 11:

> Patch Set 9: Code-Review+2
> 
> (1 comment)
> 
> Thanks for the fix!

Yunqing,
We have abstracted some functionality of av1_superres_encode into a new function: av1_realloc_and_scale_source, which re-allocates and scales the source buffer if required. We have not included the check for "av1_resize_scaled" in this function as it uses render_width and render_height to check if scaling is required. As render_width and render_height are user configurable elements, they are not good to check for resizing and thus we have kept that check specific to superres and have not abstracted it.

Patch-set: 11
Work-in-progress: false