Change to qbpm_enumerator for case STRICT_RC
Change applies only when the STRICT_RC compile flag is set. No
impact in default build.

The rate_err_tol based correction does not give a gain when using
STRICT_RC.  In fact when the overshoot and undershoot pct
parameters are set to 100% this change gives an average gain of
~0.5%.

Eg. Low res: -0.382, -0.719, -0.331 -0.528

Change-Id: I1bad84cd4b2d6b4403045e41b7faf4e440dab8cb