Update patch set 4
Patch Set 4:

Hi Yunqing,

This patch fixes the heap corruption with --row-mt=1 when resizing is enabled.

The memory related to row multi-threading was reallocated whenever there is a change in the number of tile rows/tile cols/superblock rows. However, entropy context memory 'this_tile->row_ctx' is allocated based on the number of superblock columns in a tile. Thus, whenever superblock columns in a tile change, 'this_tile->row_ctx' needs to be reallocated. In this patch, we are reallocating row multi-threading memory whenever there is a change in the number of superblock columns in addition to the existing conditions. Please note that only 'this_tile->row_ctx' needs to be reallocated whenever superblock columns change. However, we reallocate all the row multi-threading related memory (as it is not a typical use-case) in order to preserve the code simplicity.

The memory required for 'this_tile->row_ctx' in each tile can be based on superblock columns in that tile. However, we calculate the maximum of the 'number of superblock columns in a tile' over all tiles and use that maximum to allocate for each tile. We have taken the above approach as code simplicity is preserved and 'this_tile->row_ctx' is a negligible percentage (not shown in valgrind runs).

In first pass stage, 'this_tile->row_ctx' is not required. Hence, we have modified the function 'row_mt_mem_alloc' accordingly. However this leads to negligible reduction in memory (not shown in valgrind runs).

Please review.

Thanks.

Patch-set: 4
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 6245 <6245@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 13265 <13265@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 17598 <17598@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 9039 <9039@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Attention: {"person_ident":"Gerrit User 5185 \u003c5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0\u003e","operation":"ADD","reason":"Reviewer was added"}