Fix bits allocated to key-frames in single pass VBR
When the key-frame interval is configured to a very high value,
the key-frame ends up stealing bits from a very high number of
frames. This patch limits the amount of frames from which kf
can steal bits in case of single pass VBR, which helps it meet
the target bit-rate better.

Change-Id: I7ca5af11934b372477474171b23b6bbe471160c3