Fix bits allocated to key-frames in single pass VBR
When the key-frame interval is configured much higher than
the sequence length, the key-frame ends up stealing bits from
frames which will never be encoded. This patch limits the amount
of frames from which kf can steal bits in case of single pass
VBR, which helps it meet the target bit-rate better.

Change-Id: I7ca5af11934b372477474171b23b6bbe471160c3