Prevent copying already allocated metadata
The aom_copy_metadata_to_frame_buffer() function receives a buffer and a
pointer to the metadata to be copied into the buffer, but before doing
so, any previous metadata in the buffer is cleared.

Prevent the dangerous case when new metadata pointer points to the same
address than current metadata in buffer and memory is cleaned before
being assigned. Just return as success as no copy is needed.

Change-Id: I09784df3c3683e2ecaaa9e5083774c971f5902e1