Refactor fwd kf related code to prepare for kf arfs.
Previous if a frame is a keyframe and if it is a no-show frame,
it is determined as a fwd keframe. But this is not necessarily
true if we allow keyframes to be ARF_UPDATE. This patch refactors
related parts of the code to use a different strategy to determine
forward keyframes.

Change-Id: I266fbcf2ca0a463a9da7fb6059f36069e7a9047a