Use the subgop cfg to refresh the reference buffers
This will use the subgop cfg frame type rather than the update type
to decide which buffers to refresh. It uses the exact same strategy
as the current encoder default. There will be a follow up change to
add a different update strategy once this is submitted. Currently
does not work with TPL.

Change-Id: I28dc48261c7286dd85dbf56f3fb3380fe88f58e2