Update aomenc, avoid 8x8 blocks and adjust QP on motion
This CL updates aomenc to accept an optional motion data text file,
each of line of which needs to follow the following format:
"frame %d : motion flag %d\n". The aomenc program would process this
text file first to construct a record of motion flags before each run.
Before encoding each frame, the program would look into the record of
motion flags to determine if to use codec control to update the flag.

It also adds in two new features:
1. skipping 8x8 blocks during a motion
2. skipping scene detection and increase QP at the onset of a motion

The path to pass the motion flag from AppRTCMobile in WebRTC to libaom
is included in cl/319307936.

Change-Id: I5f094a4868e4de251cf569520acac0be2438218e