Add low resolution clip to ScreenContentToolsTest
Added low resolution screen content clip to ScreenContentToolsTest
to validate if screen content tools are turned on for it. Currently,
this test is turned off as it is failing. Once the issue is fixed,
it can be uncommented and turned on.

Relevant issue link: https://bugs.chromium.org/p/aomedia/issues/detail?id=2770

Change-Id: Ic9e8d24058021410e6b3a54f16aa8a1849401165