Make subgop compatible with tpl
This will pass in the gf group index from the tpl model so the subgop cfg
can be correctly indexed.

Note that the tpl model artificially extends into the next
gf group and uses defaults to approximate the behavior. Because this doesn't
use true values from the gf group, it is easier to avoid using the subgop for
this behavior.

Change-Id: Ibc2968858cb294f46fa817cb3c529cb685292f1f