Finish the wiring for extended border interintra
1. Remove the refined motion vector search. Gains appear
small and there is a bug with it (mismatch on tempete)

2. Always build a border of size 16. This is due to limitations
of the implementation and x86 SIMD instructions requiring
16-byte alignment, even with strides and offsetting.

3. Set the illum-mcomp border to 4.

4. Remove the InterPredExt struct in favor of a top-left border.
If we need the genericity of the struct (arbitrary border regions
on top/left/right/bottom) we can add it back later. For now,
I'd prefer a simpler structure for debuggin.

5. Fix bug when high-bitdepth path is forced.

Change-Id: Icc8258e56824ee1bb677c8c5e59372d44bc07504