Adjust rdmult based on tpl stats.
The current rdmult is adjusted with a static multiplier based on
the layer level of the frame, and the boost factor from
first pass stats.

We replace the adjustment strategy with tpl stats analysis,
which allows us to tune the rdmult adaptively to the content.

We accumulate the mc_dep_cost (from tpl stats) for each frame and
normalize it with respect to the GOP. Denoted as
"frame_dependency_ratio", the parameter provides an estimate of the
importance of the current frame in the GOP. For example, the ratio
of the ARF usually lies in the range of 0.2 to 0.4, meaning the ARF
accounts for 20% to 40% of information to be coded in the hiearchical
structure. The ratio is decreasing with the increase of layer level
and leaf frames usually have values around 0.05. The tpl stats based
ratio could be considered as a content adaptive estimation that is
better than a static lookup table.

We use a simple equation to determine the adjustment:
rdmult += rdmult * alpha * exp(-beta * frame_dependency_ratio).
"alpha" is a scaling factor that roughly matches the scale of the
lookup table "rd_layer_depth_factor" so that the change does not lead
to a catastrophic rate distortion surprise.
"beta" is a damping factor. Both variables are fixed for all
resolutions.

The performance change is as follows:

speed_1    ovr_psnr    ssim    vmaf_neg
lowres2    -0.127      0.594     -0.413
ugc360p    -0.350      0.315     -0.484
ugc480p    -1.080     -0.367     -1.192
hdres      -0.642      0.090     -0.311

speed_4    ovr_psnr    ssim    vmaf_neg
lowres2    -0.050      0.591     -0.480
ugc360p    -0.458      0.226     -0.694
ugc480p    -0.981     -0.248     -0.933
hdres      -0.980     -0.183     -0.405

The change also has impact on the rate accuracy. In general, we choose
"alpha" to be slightly smaller than the original scaling factor in
"rd_layer_depth_factor", and thus a smaller rdmult. Therefore the rate
optimization (Distortion + rdmult * Rate) tends to emphasize on
distortion, leading to a higher rate. It could amplify overshooting,
but also reduces undershooting.

STATS_CHANGED
Change-Id: If6589ffb00d8e6a6da57de53ba593b48048be8b8