Replace InterPredExt with border
Instead of the generic InterPredExt structure (which supports
a border region on top/left/bottom/right of any size), use a
single integer to represent the size of the top-left border.
Right now, all experiments only use the top-left, and this change
simplifies the code. If we need genericity later, we can add it
back later.

Change-Id: I6088f758f2e0553fb7eb9b2203af29aef1b47098